+++ packages/xchain-util/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Extend BaseAmount/AssetAmount to support basic arithmetic operations(add, minus, times, div)
 - Extend BaseAmount/AssetAmount to support basic comparison
+- Add type guard `isBigNumberValue` for BigNumber.Value
 
 # v.0.2.5 (2021-03-04)